Bookseller/requirements.txt
django
djangorestframework
pyyaml
requests
django-cors-headers
asgiref==3.3.1
pytz==2021.1
sqlparse==0.4.1